requirements.txt
pywebview>=4.4

